+++ examples/vpc/main.tf
@@ -1,0 +1,20 @@
+module "vpc" {
+  source = "../../modules/vpc"
+
+  vpc_cidr_block      = "10.0.0.0/16"
+  public_subnet_cidr  = "10.0.1.0/24"
+  private_subnet_cidr = "10.0.2.0/24"
+
+  az = [
+    "ap-south-1a",
+    "ap-south-1b"
+  ]
+
+  project     = "terraform-aws-platform"
+  environment = "dev"
+
+  tags = {
+    ManagedBy = "Terraform"
+    Owner     = "Platform Team"
+  }
+}